feat: add uuid validation and check for it

+++ src/services/user.service.ts
@@ -2,6 +2,7 @@ import { IncomingMessage, ServerResponse } from "http";
 import { db } from "../db/db";
 import { IUser } from "../models/users.model";
 import { IRequestWithBody } from "../models/requestWithBody";
+import { UUID_REGEXP } from "../constants";
 
 export class UserService {
     user: IUser | null;
@@ -28,8 +29,8 @@ export class UserService {
         this.res.end(JSON.stringify(data));
     };
 
-    getUserById = () => {        
-        if (this.id) {
+    getUserById = () => {     
+        if (this.id && this.id?.match(UUID_REGEXP)) {
             const data = db.getUserById(this.id);
             if (data) {
                 this.res.statusCode = 200;
@@ -38,7 +39,11 @@ export class UserService {
                 this.res.statusCode = 404;
                 this.res.end(`Wrong id. User with ${ this.id } does not exist!`)
             }
+        } else {
+            this.res.statusCode = 400;
+            this.res.end(`This id - not UUID format!`);
         }
+        
     };
 
     createNewUser = () => {
@@ -46,7 +51,7 @@ export class UserService {
             const res = db.createNewUser(this.user);
             if (res) {
                 this.res.statusCode = 200;
-                this.res.end(`New user with id ${ this.id } created!`);
+                this.res.end(`New user with id ${ this.user.id } created!`);
             } else {
                 this.res.statusCode = 404;
                 this.res.end('User with this id already exist!')
@@ -58,28 +63,34 @@ export class UserService {
     };
 
     updateUser = () => {
-        if (this.id && this.user) {
+        if (this.id && this.user && this.id?.match(UUID_REGEXP)) {
             const res = db.updateUser(this.id, this.user);
             if (res) {
                 this.res.statusCode = 200;
-                this.res.end(`User with id ${ this.id } updated!`)
+                this.res.end(`User with id ${ this.id } updated!\nUser: ${ JSON.stringify(this.user) }`)
             } else {
                 this.res.statusCode = 404; 
                 this.res.end(`User with id ${ this.id } does not exist!`)
             }
+        } else {
+            this.res.statusCode = 400;
+            this.res.end(`This id - not UUID format!`);
         }
     };
 
     deleteUser = () => {
-        if (this.id) {
+        if (this.id && this.id?.match(UUID_REGEXP)) {
             const res = db.deleteUser(this.id)
             if (res) {
-                this.res.statusCode = 204;
                 this.res.end(`User with id ${ this.id } deleted!`);
+                this.res.statusCode = 204;
             } else {
                 this.res.statusCode = 404;
                 this.res.end(`User with id ${ this.id } not found!`);
             }
+        } else {
+            this.res.statusCode = 400;
+            this.res.end(`This id - not UUID format!`);
         }
     };
 }
